changed most of the shortcuts exec to thunks, changes newThought util to thunk
