+++ src/shortcuts/cursorNext.js
@@ -1,15 +1,8 @@
 import React from 'react'
 import { store } from '../store.js'
 
-// util
-import {
-  contextOf,
-  getThoughtAfter,
-  headValue,
-  isDivider,
-  nextThoughtElement,
-  restoreSelection,
-} from '../util.js'
+// action-creators
+import { cursorNext } from '../action-creators/cursorNext'
 
 const Icon = ({ fill = 'black', size = 20, style }) => <svg version="1.1" className="icon" xmlns="http://www.w3.org/2000/svg" width={size} height={size} fill={fill} style={style} viewBox="0 0 19.481 19.481" enableBackground="new 0 0 19.481 19.481">
   <g>
@@ -17,48 +10,11 @@ const Icon = ({ fill = 'black', size = 20, style }) => <svg version="1.1" classN
   </g>
 </svg>
 
-const execThunk = (e) => (dispatch, getState) => {
-  const { cursor } = getState()
-
-  // select next editable
-  if (cursor) {
-    const next = nextThoughtElement(cursor)
-    if (next) {
-      const editable = next.querySelector('.editable')
-
-      if (editable) {
-        // editable focus does not work when moving from a divider for some reason
-        if (isDivider(headValue(cursor))) {
-          const nextThought = getThoughtAfter(cursor)
-          const nextThoughtsRanked = contextOf(cursor).concat(nextThought)
-          restoreSelection(nextThoughtsRanked)
-        }
-        else {
-          editable.focus()
-        }
-      }
-      else if (next.querySelector('.divider')) {
-        const nextThought = getThoughtAfter(cursor)
-        const nextThoughtsRanked = contextOf(cursor).concat(nextThought)
-        dispatch({ type: 'setCursor', thoughtsRanked: nextThoughtsRanked })
-        document.getSelection().removeAllRanges()
-      }
-    }
-  }
-  // if no cursor, select first editable
-  else {
-    const firstEditable = document.querySelector('.editable')
-    if (firstEditable) {
-      firstEditable.focus()
-    }
-  }
-}
-
 export default {
   id: 'cursorNext',
   name: 'Cursor Next Thought',
   description: 'Move the cursor to the next thought, skipping expanded children.',
   keyboard: { key: 'ArrowDown', meta: true },
   svg: Icon,
-  exec: e => store.dispatch(execThunk(e))
+  exec: e => store.dispatch(cursorNext(e))
 }
