+++ src/shortcuts/moveThoughtDown.js
@@ -1,16 +1,8 @@
 import React from 'react'
 import { store } from '../store.js'
 
-// util
-import {
-  contextOf,
-  headRank,
-  headValue,
-  getRankAfter,
-  nextSibling,
-  restoreSelection,
-  rootedContextOf,
-} from '../util.js'
+// action-creators
+import { moveThoughtDown } from '../action-creators/moveThoughtDown'
 
 const Icon = ({ fill = 'black', size = 20, style }) => <svg version="1.1" className="icon" xmlns="http://www.w3.org/2000/svg" width={size} height={size} fill={fill} style={style} viewBox="0 0 19.481 19.481" enableBackground="new 0 0 19.481 19.481">
   <g>
@@ -19,40 +11,10 @@ const Icon = ({ fill = 'black', size = 20, style }) => <svg version="1.1" classN
 </svg>
 
 export default {
-    id: 'moveThoughtDown',
-    name: 'Move Thought Down',
-    description: 'Move the current thought down.',
-    keyboard: { key: 'ArrowDown', meta: true, shift: true },
-    svg: Icon,
-    exec: () => {
-      const { cursor } = store.getState()
-
-      if (cursor) {
-
-        const context = contextOf(cursor)
-        const value = headValue(cursor)
-        const rank = headRank(cursor)
-
-        const nextThought = nextSibling(value, rootedContextOf(cursor), rank)
-        if (nextThought) {
-
-          // store selection offset before existingThoughtMove is dispatched
-          const offset = window.getSelection().focusOffset
-
-          const rankNew = getRankAfter(context.concat(nextThought))
-          const newPath = context.concat({
-            value,
-            rank: rankNew
-          })
-
-          store.dispatch({
-            type: 'existingThoughtMove',
-            oldPath: cursor,
-            newPath
-          })
-
-          restoreSelection(newPath, { offset })
-        }
-      }
-    }
-  }
+  id: 'moveThoughtDown',
+  name: 'Move Thought Down',
+  description: 'Move the current thought down.',
+  keyboard: { key: 'ArrowDown', meta: true, shift: true },
+  svg: Icon,
+  exec: () => store.dispatch(moveThoughtDown())
+}
