+++ src/shortcuts/moveThoughtUp.js
@@ -1,16 +1,8 @@
 import React from 'react'
 import { store } from '../store.js'
 
-// util
-import {
-  contextOf,
-  headRank,
-  headValue,
-  getRankBefore,
-  prevSibling,
-  restoreSelection,
-  rootedContextOf,
-} from '../util.js'
+// action-creators
+import { moveThoughtUp } from '../action-creators/moveThoughtUp'
 
 const Icon = ({ fill = 'black', size = 20, style }) => <svg version="1.1" className="icon" xmlns="http://www.w3.org/2000/svg" width={size} height={size} fill={fill} style={style} viewBox="0 0 19.481 19.481" enableBackground="new 0 0 19.481 19.481">
   <g>
@@ -19,40 +11,10 @@ const Icon = ({ fill = 'black', size = 20, style }) => <svg version="1.1" classN
 </svg>
 
 export default {
-    id: 'moveThoughtUp',
-    name: 'Move Thought Up',
-    description: 'Move the current thought up.',
-    keyboard: { key: 'ArrowUp', meta: true, shift: true },
-    svg: Icon,
-    exec: () => {
-      const { cursor } = store.getState()
-
-      if (cursor) {
-
-        const context = contextOf(cursor)
-        const value = headValue(cursor)
-        const rank = headRank(cursor)
-
-        const prevThought = prevSibling(value, rootedContextOf(cursor), rank)
-        if (prevThought) {
-
-          // store selection offset before existingThoughtMove is dispatched
-          const offset = window.getSelection().focusOffset
-
-          const rankNew = getRankBefore(context.concat(prevThought))
-          const newPath = context.concat({
-            value,
-            rank: rankNew
-          })
-
-          store.dispatch({
-            type: 'existingThoughtMove',
-            oldPath: cursor,
-            newPath
-          })
-
-          restoreSelection(newPath, { offset })
-        }
-      }
-    }
-  }
+  id: 'moveThoughtUp',
+  name: 'Move Thought Up',
+  description: 'Move the current thought up.',
+  keyboard: { key: 'ArrowUp', meta: true, shift: true },
+  svg: Icon,
+  exec: () => store.dispatch(moveThoughtUp())
+}
