+++ src/shortcuts/newThoughtAbove.js
@@ -1,9 +1,10 @@
 import React from 'react'
+import { store } from '../store'
 
-// util
+// action-creators
 import {
   newThought,
-} from '../util.js'
+} from '../action-creators/newThought'
 
 const Icon = ({ fill = 'black', size = 20, style }) => <svg version="1.1" className="icon" xmlns="http://www.w3.org/2000/svg" width={size} height={size} fill={fill} style={style} viewBox="0 0 19.481 19.481" enableBackground="new 0 0 19.481 19.481">
   <g>
@@ -18,7 +19,5 @@ export default {
   gesture: 'rul',
   keyboard: { key: 'Enter', shift: true },
   svg: Icon,
-  exec: () => {
-    newThought({ insertBefore: true })
-  }
+  exec: () => store.dispatch(newThought({ insertBefore: true }))
 }
