+++ src/shortcuts/newUncle.js
@@ -1,10 +1,12 @@
 import React from 'react'
 import { store } from '../store.js'
 
+// action-creators
+import { newThought } from '../action-creators/newThought'
+
 // util
 import {
   contextOf,
-  newThought,
 } from '../util.js'
 
 const Icon = ({ fill = 'black', size = 20, style }) => <svg version="1.1" className="icon" xmlns="http://www.w3.org/2000/svg" width={size} height={size} fill={fill} style={style} viewBox="0 0 19.481 19.481" enableBackground="new 0 0 19.481 19.481">
@@ -24,9 +26,9 @@ export default {
   exec: () => {
     const { cursor } = store.getState()
     if (cursor && cursor.length > 1) {
-      newThought({
+      store.dispatch(newThought({
         at: contextOf(cursor)
-      })
+      }))
     }
   }
 }
